+++ starsim/timepars.py
